+++ packages/manager/apps/sms/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.6.23](https://github.com/ovh/manager/compare/@ovh-ux/manager-sms-app@7.6.22...@ovh-ux/manager-sms-app@7.6.23) (2023-02-23)
+
+**Note:** Version bump only for package @ovh-ux/manager-sms-app
+
+
+
+
+
 ## [7.6.22](https://github.com/ovh/manager/compare/@ovh-ux/manager-sms-app@7.6.21...@ovh-ux/manager-sms-app@7.6.22) (2023-02-20)
 
 **Note:** Version bump only for package @ovh-ux/manager-sms-app